Add system bar
