+++ src/components/CodeMirror.vue
@@ -67,6 +67,7 @@ export default Vue.extend({
     onMounted(() => {
       state.data = JSON.stringify(props.data, null, '\t');
       window.addEventListener('resize', resizeCodeMirror);
+      setTimeout(resizeCodeMirror, 1);
     });
 
     watch(
@@ -83,14 +84,12 @@ export default Vue.extend({
 
 <style>
 .codeMirror {
-  overflow: hidden;
+  height: calc(100vh - 48px - 24px);
 }
 
-.CodeMirror {
-  border: 1px solid #eee;
+.codeMirror .CodeMirror {
+  border: 1px solid rgba(0, 0, 0, 0.12);
   font-family: monospace;
   font-size: 12px;
-  width: 512px;
-  height: calc(100vh - 48px);
 }
 </style>
