+++ src/components/app/AppDetails.vue
